Correct invalid .obj files
There were a few obj files that contained some instructions which
obj2gtfl was not able to recognize, which then through off the vertex
indexing and caused the conversion to fail with an error:
"Normal index ___ is out of bound"

- hs903fr5133
- kn073qt9437
- pr756nc2659
- vr966cz0421
- vx600rw8369
- yg363gd9984

I found some information about it at https://github.com/CesiumGS/obj2gltf/issues/243

To remediate the converter first creates a copy of the original .obj
file, and edits "nan" values to "0.0" in the copy. This copy is then
used to run the obj2gltf program.